changed size_t to uint16_t (fix for #398)
git-svn-id: http://alma-dms/svn/sakura/trunk@752 9a79bd28-2a0b-46f8-beb7-14fc758f186d

+++ libsakura/src/baseline.cc
@@ -276,7 +276,7 @@ inline void SubtractBaseline(size_t num_data, float const *data_arg,
 		num_fitting_max = 1;
 	}
 
-	for (size_t i = 0; i < num_fitting_max; ++i) {
+	for (uint16_t i = 0; i < num_fitting_max; ++i) {
 		if (i == 0) {
 			for (size_t j = 0; j < num_data; ++j) {
 				final_mask[j] = mask[j];
